async_reports/requirement.txt
Django==1.4
HTSQL==2.3.0
PyHamcrest==1.6
PyYAML==3.10
Pygments==1.5
amqplib==1.0.2
anyjson==0.3.1
bpython==0.11
celery==2.5.3
distribute==0.6.26
django-nose==1.0
kombu==2.1.8
nose==1.1.2
python-dateutil==1.5
wsgiref==0.1.2
pip install django-async-messages
django-jsonfield==0.8.10
